Moved out OutdoorTerrain
