+++ src/Engine/Graphics/Outdoor.h
@@ -16,6 +16,7 @@
 #include "LocationInfo.h"
 #include "LocationTime.h"
 #include "LocationFunctions.h"
+#include "OutdoorTerrain.h"
 
 struct DecalBuilder;
 struct SpellFxRenderer;
@@ -23,23 +24,9 @@ struct TileData;
 struct RenderVertexSoft;
 struct ODMRenderParams;
 
-struct OutdoorLocationTileType {
-    Tileset tileset = Tileset_NULL;
-    uint16_t uTileID = 0;
-};
-
 struct DMap {
     uint8_t field0;
     uint8_t field1;
-};
-
-struct OutdoorLocationTerrain {
-    bool ZeroLandscape();
-
-    std::array<OutdoorLocationTileType, 4> pTileTypes;  // [3] is road tileset.
-    std::array<uint8_t, 128 * 128> pHeightmap = {};
-    std::array<uint8_t, 128 * 128> pTilemap = {};
-    std::array<uint8_t, 128 * 128> pAttributemap = {};
 };
 
 struct OutdoorLocation {
@@ -125,7 +112,7 @@ struct OutdoorLocation {
     std::string location_filename;
     std::string location_file_description;
     std::string sky_texture_filename;
-    OutdoorLocationTerrain pTerrain;
+    OutdoorTerrain pTerrain;
     std::vector<BSPModel> pBModels;
     std::vector<Pid> pFaceIDLIST;
     std::array<uint32_t, 128 * 128> pOMAP;
