+++ src/Engine/Graphics/OutdoorTerrain.cpp
@@ -1,0 +1,9 @@
+#include "OutdoorTerrain.h"
+
+//----- (0047CCE2) --------------------------------------------------------
+bool OutdoorTerrain::ZeroLandscape() {
+    this->pHeightmap.fill(0);
+    this->pTilemap.fill(90);
+    this->pAttributemap.fill(0);
+    return true;
+}
